Implement file upload

+++ includes/forms/class-user-delete.php
@@ -44,15 +44,17 @@ class User_Delete extends Form {
 	public function submit() {
 		parent::submit();
 
-		require_once ABSPATH . 'wp-admin/includes/user.php';
+		if ( is_user_logged_in() ) {
+			require_once ABSPATH . 'wp-admin/includes/user.php';
 
-		// Check password.
-		if ( ! wp_check_password( $this->get_value( 'password' ), wp_get_current_user()->user_pass, get_current_user_id() ) ) {
-			$this->errors[] = esc_html__( 'Password is incorrect.', 'hivepress' );
-		} elseif ( ! current_user_can( 'manage_options' ) ) {
+			// Check password.
+			if ( ! wp_check_password( $this->get_value( 'password' ), wp_get_current_user()->user_pass, get_current_user_id() ) ) {
+				$this->errors[] = esc_html__( 'Password is incorrect.', 'hivepress' );
+			} elseif ( ! current_user_can( 'manage_options' ) ) {
 
-			// Delete user.
-			wp_delete_user( get_current_user_id() );
+				// Delete user.
+				wp_delete_user( get_current_user_id() );
+			}
 		}
 
 		return empty( $this->errors );
